requirements.txt
fastapi==0.110.1
Requests==2.31.0
opencv-python==4.9.0.80
python-multipart==0.0.9
uvicorn==0.29.0
python-dotenv==1.0.1
openai==1.17.0